sweet alert utilized for confirm/cancel of delete feedback

+++ src/components/Admin/Admin.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import axios from 'axios';
 // import { confirmAlert } from 'react-confirm-alert';
 // import 'react-confirm-alert/src/react-confirm-alert.css';
+import swal from 'sweetalert';
 
 class Admin extends Component {
 
@@ -28,39 +29,55 @@ class Admin extends Component {
     this.getFeedback();
   }
 
-  // submit = () => {
-  //   console.log('delete button clicked');
-  //   confirmAlert({
-  //     title: 'Confirm to submit',
-  //     message: 'Are you sure to do this.',
-  //     buttons: [
-  //       {
-  //         label: 'Yes',
-  //         onClick: () => alert('Click Yes')
-  //       },
-  //       {
-  //         label: 'No',
-  //         onClick: () => alert('Click No')
-  //       }
-  //     ]
-  //   })
-  // };
-
   // delete feedback from database
-  deleteFeedback = (id) => {
-    console.log('row id to delete:', id);
-    console.log('in deleteFeedback');
-    axios({
-        method: 'DELETE',
-        url: `/feedback/${id}`
-    })
-    .then((response) => {
-        this.getFeedback();
-    })
-    .catch((error) => {
-        console.log('DELETE error:', error)
-    })
-  }
+  // deleteFeedback = (id) => {
+  //   console.log('row id to delete:', id);
+  //   console.log('in deleteFeedback');
+  //   axios({
+  //       method: 'DELETE',
+  //       url: `/feedback/${id}`
+  //   })
+  //   .then((response) => {
+  //       this.getFeedback();
+  //   })
+  //   .catch((error) => {
+  //       console.log('DELETE error:', error)
+  //   })
+  // }
+
+  // DELETE from db - used sweetalert to confirm/cancel delete
+  deleteFeedbackSwal = (id) => {
+    swal({
+      title: 'Are you sure?',
+      text: 'Once deleted, the feedback will be removed permanently!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+      })
+      .then((willDelete) => {
+      if (willDelete) {
+          swal('The feedback has been deleted!', {
+          icon: 'success',
+          });
+      console.log('in deleteTask');
+      axios({
+          method: 'DELETE',
+          url: `/feedback/${id}`
+      })
+      .then((response) => {
+          console.log(response);
+          this.getFeedback();
+      })
+      .catch((error) => {
+          console.log(error);        
+      })
+      } // end if 
+      else {
+          swal('The feedback remains!');
+      } // end else
+     a});
+  } // end deleteFeedbackSwal
+
 
   render() {
     return (
@@ -83,7 +100,7 @@ class Admin extends Component {
                         <td>{submission.understanding}</td>
                         <td>{submission.support}</td>
                         <td>{submission.comments}</td>
-                        <td><button onClick={() => this.deleteFeedback(submission.id)}>DELETE</button></td>
+                        <td><button onClick={() => this.deleteFeedbackSwal(submission.id)}>DELETE</button></td>
                     </tr>
                   )}
                 </tbody>
